+++ deptrac.yaml
@@ -6,8 +6,6 @@ parameters:
     - name: Legacy
       collectors:
         - type: className
-          regex: ^Bavix\\.*Legacy$
-        - type: className
           regex: ^Bavix\\.*\\Objects\\Cart$
 
     - name: Contract
@@ -242,11 +240,8 @@ parameters:
       - Dto
 
     Legacy:
-      - AssemblerDtoInterface
       - InternalException
       - ServiceInterface
       - DtoInterface
-      - UIException
       - Contract
-      - Model
       - Dto # Cart from objects
