+++ src/Services/TransactionService.php
@@ -1,0 +1,76 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bavix\Wallet\Services;
+
+use Bavix\Wallet\Interfaces\Wallet;
+use Bavix\Wallet\Internal\Dto\TransactionDtoInterface;
+use Bavix\Wallet\Internal\Exceptions\LockProviderNotFoundException;
+use Bavix\Wallet\Internal\Exceptions\RecordNotFoundException;
+use Bavix\Wallet\Models\Transaction;
+
+final class TransactionService implements TransactionServiceInterface
+{
+    public function __construct(
+        private AssistantServiceInterface $assistantService,
+        private RegulatorServiceInterface $regulatorService,
+        private PrepareServiceInterface $prepareService,
+        private CastServiceInterface $castService,
+        private AtmServiceInterface $atmService,
+    ) {
+    }
+
+    /**
+     * @throws LockProviderNotFoundException
+     * @throws RecordNotFoundException
+     */
+    public function makeOne(
+        Wallet $wallet,
+        string $type,
+        float|int|string $amount,
+        ?array $meta,
+        bool $confirmed = true
+    ): Transaction {
+        assert(in_array($type, [Transaction::TYPE_DEPOSIT, Transaction::TYPE_WITHDRAW], true));
+
+        if ($type === Transaction::TYPE_DEPOSIT) {
+            $dto = $this->prepareService->deposit($wallet, (string) $amount, $meta, $confirmed);
+        } else {
+            $dto = $this->prepareService->withdraw($wallet, (string) $amount, $meta, $confirmed);
+        }
+
+        $transactions = $this->apply([
+            $dto->getWalletId() => $wallet,
+        ], [$dto]);
+
+        return current($transactions);
+    }
+
+    /**
+     * @param non-empty-array<int|string, Wallet>           $wallets
+     * @param non-empty-array<int, TransactionDtoInterface> $objects
+     *
+     * @throws LockProviderNotFoundException
+     * @throws RecordNotFoundException
+     *
+     * @return non-empty-array<string, Transaction>
+     */
+    public function apply(array $wallets, array $objects): array
+    {
+        $transactions = $this->atmService->makeTransactions($objects); // q1
+        $totals = $this->assistantService->getSums($objects);
+
+        foreach ($totals as $walletId => $total) {
+            $wallet = $wallets[$walletId] ?? null;
+            assert($wallet !== null);
+
+            $object = $this->castService->getWallet($wallet);
+            assert((int) $object->getKey() === $walletId);
+
+            $this->regulatorService->increase($object, $total);
+        }
+
+        return $transactions;
+    }
+}
