+++ src/Services/TransactionServiceInterface.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bavix\Wallet\Services;
+
+use Bavix\Wallet\Interfaces\Wallet;
+use Bavix\Wallet\Internal\Dto\TransactionDtoInterface;
+use Bavix\Wallet\Internal\Exceptions\LockProviderNotFoundException;
+use Bavix\Wallet\Internal\Exceptions\RecordNotFoundException;
+use Bavix\Wallet\Models\Transaction;
+
+interface TransactionServiceInterface
+{
+    /**
+     * @throws LockProviderNotFoundException
+     * @throws RecordNotFoundException
+     */
+    public function makeOne(
+        Wallet $wallet,
+        string $type,
+        float|int|string $amount,
+        ?array $meta,
+        bool $confirmed = true
+    ): Transaction;
+
+    /**
+     * @param non-empty-array<int|string, Wallet>           $wallets
+     * @param non-empty-array<int, TransactionDtoInterface> $objects
+     *
+     * @throws LockProviderNotFoundException
+     * @throws RecordNotFoundException
+     *
+     * @return non-empty-array<string, Transaction>
+     */
+    public function apply(array $wallets, array $objects): array;
+}
